+++ tests/test_main.py
@@ -3,11 +3,13 @@ import sys
 from pathlib import Path
 
 import imgcompare
+import pytest
 from nocairosvg import svg2bitmap
 
 THISDIR = str(Path(__file__).resolve().parent)
 sys.path.insert(0, str(Path(THISDIR).parent))
 logoFile = f"{THISDIR}/data/logo"
+notExistsFile = f"{THISDIR}/data/notExists"
 
 from svgtrace import trace
 
@@ -28,3 +30,8 @@ def test_bw():
 def test_colour():
 	Path(f"{logoFile}.svg").write_text(trace(f"{logoFile}.png"), encoding="utf-8")
 	aux_comparesvg(f"{logoFile}.svg", "logo-actual.png")
+
+
+def test_notExists():
+	with pytest.raises(FileNotFoundError):
+		Path(f"{notExistsFile}.svg").write_text(trace(f"{notExistsFile}.png"), encoding="utf-8")